auth-service/requirements.txt
Django==5.2.12
djangorestframework==3.15.2
PyJWT==2.9.0
psycopg2-binary==2.9.9
